处理市场比较法得一些bug以及调整收益法调用市场比较法得bug

+++ assess-app/src/main/java/com/copower/pmcc/assess/dto/input/method/MarketCompareItemDto.java
@@ -8,7 +8,7 @@ import java.math.BigDecimal;
 public class MarketCompareItemDto {
     private String name;
     private String value;
-    private Integer score;
+    private BigDecimal score;
     private BigDecimal ratio;
 
 
@@ -28,11 +28,11 @@ public class MarketCompareItemDto {
         this.value = value;
     }
 
-    public Integer getScore() {
+    public BigDecimal getScore() {
         return score;
     }
 
-    public void setScore(Integer score) {
+    public void setScore(BigDecimal score) {
         this.score = score;
     }
